backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
playwright
pydantic
requests
python-dotenv
apscheduler
tzlocal
